Add basic MSON Refract tests
This removes the old MSON AST based tests in favor of new MSON Refract ones.
The tests are significantly less complex because the exact structure of the
MSON result should be tested in the upstream library, not the language binding.
Here we just test for a very basic structure in both the attributes and
data structures sections.